created python file for function specific to scp
